+++ cloudsmith/CHANGELOG.md
@@ -5,7 +5,7 @@
 ***Added***:
 
 * Metric support for storage usage, bandwidth usage, token usage, and member activity
-* Event support for security vulnerability scan results, audit logs, policy violations related to vulnerabilities and license compliance
+* Event support for security vulnerability scan results, audit logs, policy violations related to vulnerabilities, and license compliance
 
 ### 0.0.2 / 2021-09-08